livekit_agent/requirements.txt
livekit-agents[silero,turn-detector,openai,deepgram,elevenlabs,groq,langchain]~=1.4
livekit-plugins-noise-cancellation~=0.2
python-dotenv
langchain-openai
langchain-tavily